requirements.txt
alembic==1.0.11
amqp==2.5.1
asn1crypto==0.24.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
blinker==1.4
celery==4.4.0rc1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
elasticsearch==7.0.2
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.3
Mako==1.0.10
MarkupSafe==1.1.1
pycosat==0.6.3
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyperclip==1.7.0
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
redis==3.2.1
requests==2.22.0
ruamel-yaml==0.15.46
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.7
urllib3==1.24.2
vine==1.3.0
Werkzeug==0.15.5
WTForms==2.2.1
